+++ packages/cli/src/codemods/transforms/transform-modules-to-packages-v33/transform-modules-to-packages-v33.ts
@@ -1,10 +1,22 @@
 import { type AstCliContext, type AstTransform } from '@ag-grid-devtools/ast';
 
 import { jsCodeShiftTransform } from '../../plugins/jscodeshift';
-import { processImports } from './transformers/module-imports';
+import { addAllCommunityModule } from './transformers/add-all-community';
+import { chartImports } from './transformers/chart-imports';
+import { updateDeprecatedModules } from './transformers/deprecated-modules';
+import { registerModule } from './transformers/register-module';
+import { updateImportPaths } from './transformers/update-import-paths';
+import { updateStyles } from './transformers/update-styles';
 
 const transform: AstTransform<AstCliContext> = function transformModulesToPackagesV33(babel) {
-  return jsCodeShiftTransform(processImports)(babel);
+  return jsCodeShiftTransform(
+    registerModule,
+    updateStyles,
+    chartImports,
+    updateImportPaths,
+    addAllCommunityModule,
+    updateDeprecatedModules,
+  )(babel);
 };
 
 export default transform;
